Fix warning for double promotion in XmlAttribute
This one should be converting to double in the first place.

+++ src/Xml/XmlAttribute.cpp
@@ -175,7 +175,7 @@ int XmlAttribute::intValue() const
  */
 double XmlAttribute::doubleValue() const
 {
-	return std::stof(_value);
+	return std::stod(_value);
 }
